requirements.txt
# Travel App Dependencies

# Core CLI framework
typer[all]>=0.9.0

# Environment configuration
python-dotenv>=1.0.0

# HTTP requests for API calls
requests>=2.31.0

# Data formatting and display
tabulate>=0.9.0

# Data validation and parsing
pydantic>=2.0.0

# Web UI framework
streamlit>=1.28.0

# Data manipulation for web UI
pandas>=2.0.0

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0

# Timezone handling (Python 3.9+ has zoneinfo built-in)
# backports.zoneinfo>=0.2.1; python_version<"3.9"
